refactor: th new building builder class is done
