+++ test/models/building/test_item_producing.py
@@ -1,6 +1,7 @@
 from django.test import TestCase
 from django.contrib.auth.models import User
-from app.models import Landscape, Land, Company, Building, BuildingType, Item
+from app.models import Landscape, Land, Company, BuildingType, Item
+from app.core.services.builders.building_builder import BuildingBuilder
 
 
 class BuildingItemProducingTestCase(TestCase):
@@ -15,8 +16,8 @@ class BuildingItemProducingTestCase(TestCase):
         self.land.purchase_landscape(self.company)
         building: BuildingType = BuildingType.objects.get_building_by_type('supreme mine')
         self.company.balance = building.get_rent_cost(0)
-        self.building = Building.objects.create_building("supreme mine", 'mymine', self.company,
-                                                         'rent', 0, self.land)
+        self.building = BuildingBuilder.construct("supreme mine", 'mymine', self.company,
+                                                  'rent', 0, self.land)
     
     def test_building_can_produce(self):
         item = Item.objects.get(name='diamond')  # assuming diamond is produced from surpeme mine
