+++ test/models/company/test_company_agent.py
@@ -1,7 +1,8 @@
 from django.test import TestCase
 from django.contrib.auth.models import User
-from app.models import Landscape, Land, Company, Building, BuildingType, Place
+from app.models import Landscape, Land, Company, BuildingType, Place
 from app.core.services.builders.agent_builder import AgentBuilder
+from app.core.services.builders.building_builder import BuildingBuilder
 
 
 class CompanyAgent(TestCase):
@@ -27,8 +28,8 @@ class CompanyAgent(TestCase):
     def purchase_building(self):
         mine: BuildingType = BuildingType.objects.get_building_by_type('supreme mine')
         self.company.balance = mine.get_buy_cost()
-        building = Building.objects.create_building(mine.name, 'myfirstbuilding', self.company,
-                                                    'buy', 0, self.land)
+        building = BuildingBuilder.construct(mine.name, 'myfirstbuilding', self.company,
+                                             'buy', 0, self.land)
         return building
     
     def create_agent(self):
